requirements.txt
click==7.1.2
pandas==1.0.4
nltk==3.5
scikit-learn==0.23.1
# FIXME: Locking werkzeug until the resolution of pallets/werkzeug#1714
werkzeug==1.0.1
gensim==3.8.3
PyYAML==5.3.1
spacy==2.3.0